create two files for each placement delay model
